requirements.txt
Django>=3.2,<5.0
djangorestframework>=3.12
django-filter>=2.4
psycopg2-binary>=2.8
gunicorn
python-dotenv
dj-database-url
django-seed
django-cors-headers
whitenoise
pandas